Config, random posts, loop subs in order
I added config file for inputs for list of subreddits to choose from. I
also added a counter in that file that loops across the subreddits in
order, not randomly so that it doesn't pick the same sub/post. I added a
random way to pick posts from the top 50 in each subreddit for more
variety.